Project Code
River level prediction using XGBoost and Random Forest. 
1. The raw data is pre-processed using Athena. 
2. GAN data is generated using the processed data. 
3. Features are generated which are passed to Random Forest and XGBoost which in turn predict the river level. 
4. manifest.json are uploaded to QuickSight for visualization. 
5. Flood Triggering comparison is done between both algorithms. Effect of time-period of the data on accuracy is also assessed. 

ACCURACY WAS INVERSELY PROPORTIONAL TO THE TIME PERIOD OF THE DATA. WITH AN INCREASING TIME-PERIOD, THE NUMBER OF ROWS TO DESCRIBE FOR A GIVEN TIMELINE DECREASES. 

15-minute and 1-hour time-period were considered. If we consider 2 hours of data, the 15-minute time-period dataset will have 8 rows of data while the 1-hour time-period dataset would have just 2 rows. Hence, with more data to describe a situation accuracy increases. 

XGBoost was found to be more accurate and precise in prediction than Random Forest in either of the time-period datasets.